requirements.txt
aiogram==3.0.0b7
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0